+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 # main
 
 - Manually generate spacing for description in the help dialogue to replaces tabs and avoid visual errors
+- ⚠️ Move int Flag.type to bool Flag.value: name and data type change in struct Flag
 
 # 0.0.4